reuse default search ordering for alert samples
